Move picasso and misc fixes
