+++ app/src/main/java/com/github/gotify/service/WebSocketService.java
@@ -18,7 +18,6 @@ import androidx.core.app.NotificationCompat;
 import androidx.core.content.ContextCompat;
 import com.github.gotify.MissedMessageUtil;
 import com.github.gotify.NotificationSupport;
-import com.github.gotify.PicassoHandler;
 import com.github.gotify.R;
 import com.github.gotify.Settings;
 import com.github.gotify.Utils;
@@ -30,6 +29,7 @@ import com.github.gotify.log.Log;
 import com.github.gotify.log.UncaughtExceptionHandler;
 import com.github.gotify.messages.Extras;
 import com.github.gotify.messages.MessagesActivity;
+import com.github.gotify.picasso.PicassoHandler;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.TimeUnit;
